refactor: don't import Style from .core
